fix: move window objects into their own respective headers
